ABCCat

Description

Provide an representation of categories of the ABC Grammar.

Types

data ABCCat

ABCCat represents ABC categories.

Constructors

Bottom
The Bottom ⊥.

BaseCategory
An atomic catetory.

Fields

name :: Text

The name of the category.

LeftFunctor
A functor category X\C which takes an X to its left as its argument.

Fields

antecedent :: ABCCat

The argument, corresponding to the X above.

consequence :: ABCCat

The base, corresponding to the C above.

RightFunctor
A functor category C/X which takes an X to its right as its argument.

Fields

antecedent :: ABCCat

The argument, corresponding to the X above.

consequence :: ABCCat

The base, corresponding to the C above.

Bundled Patterns

pattern (:\:) :: ABCCat -> ABCCat -> ABCCat infixl 9
Compose two categories to form a left functor category.

Example:

>>> (BaseCategory "NP") :\: (BaseCategory "S")
LeftFunctor (BaseCategory "NP") (BaseCategory "S")

pattern (:/:) :: ABCCat -> ABCCat -> ABCCat infixl 9
Compose two categories to form a right functor category.

Example:

>>> (BaseCategory "VP") :/: (BaseCategory "NP")
RightFunctor (BaseCategory "NP") (BaseCategory "VP")

data ABCStatusFC

ABCStatusFC represents the rule which has been used in a process of reduction.

Constructors

FCLeft Int
Indicating that a left reduction/functional-composition rule is used.

FCRight Int
Indicating that a right reduction/functional-composition rule is used.

Failed
Indicating that the reduction cannot be achieved.

Patterns

Constants

strBot :: Text

The text representation of the bottom.

Functions

(<^#>) infix 6

Arguments

:: ABCCat
The category to the left.

-> ABCCat
The category to the right.

-> (ABCStatusFC, ABCCat)
The result.

Taking two ABC categories, try a reduction from them, leaving the trace of the used rule.

(<^>) infixr 6

Arguments

:: ABCCat
The category to the left.

-> ABCCat
The category to the right.

-> ABCCat
The result.

Taking two ABC categories, try a reduction from them.