ABCCategory

Description

Provide an representation of categories of the ABC Grammar. The parser is available at the Parser module.

Types

data ABCCategory

ABCCategory represents ABC categories.

Constructors

Bottom
The Bottom ⊥.

BaseCategory
An atomic catetory.

Fields

name :: Text

The name of the category.

LeftFunctor
A functor category X\C which takes an X to its left as its argument.

Fields

antecedent :: ABCCategory

The argument, corresponding to the X above.

consequence :: ABCCategory

The base, corresponding to the C above.

RightFunctor
A functor category C/X which takes an X to its right as its argument.

Fields

antecedent :: ABCCategory

The argument, corresponding to the X above.

consequence :: ABCCategory

The base, corresponding to the C above.

Subsidiaries

data ABCComment a

Constructors

ABCComment

Fields

content :: a

comment :: Text

type ABCCategoryCommented = ABCComment ABCCategory

data ABCStatusFC

Constructors

FCLeft Int

FCRight Int

Failed

Constants

strBot :: Text

The text representation of the bottom.

Functions

For Composition

(</>)

Arguments

:: ABCCategory
The base category C

-> ABCCategory
The argument category X.

-> ABCCategory
The resulted category C/X.

Compose two categories to form a right functor category.

Example:

>>> (BaseCategory "VP") </> (BaseCategory "NP")
RightFunctor (BaseCategory "NP") (BaseCategory "VP")

makeLeftAdjunct :: ABCCategory -> ABCCategory

(<\>)

Arguments

:: ABCCategory
The argument category X.

-> ABCCategory
The base category C

-> ABCCategory
The resulted category X\C.

Compose two categories to form a left functor category.

Example:

>>> (BaseCategory "NP") <\> (BaseCategory "S")
LeftFunctor (BaseCategory "NP") (BaseCategory "S")

makeRightAdjunct :: ABCCategory -> ABCCategory

For Reduction

(<^>) :: ABCCategory -> ABCCategory -> ABCCategory

reduceWithResult :: ABCCategory -> ABCCategory -> (ABCCategory, ABCStatusFC)

reduceWithLog :: ABCCategory -> ABCCategory -> ABCCategoryCommented