CatPlus

Description

Provide an representation of categories of the Kainoki Treebank. The parser is available at the Parser module.

Types

data CatPlus cat

A record structure that represents (Keyaki) tree node labels which are annotated with our own additional attributes.

Constructors

Term
A terminal node label, which contains nothing other than a word.

Fields

word :: Text

A word.

NonTerm
A non-terminal node label comprised of a (Keyaki or ABC) category and extra annotations for the ABC relabeling.

Fields

cat :: cat

A (Keyaki / ABC) category.

index :: Maybe Int

An ICH index on overt parts of ICH configufations.

role :: DepMarking

A grammatical role in the ABC Grammar.

deriv :: Text

The special ABC derivational rule (e.g. unary rules), if any, invoked by this node. Empty as default.

scope :: [Int]

To be abolished.

covertArgs :: [(Int, cat)]

To be abolished

A list of covert arguments (*pro* in Keyaki), each specified with its scopal rank in the clause and its category.

attrs :: Map Text Text

Other attributes.

Bundled Patterns

pattern (:#:) :: cat -> (cat -> CatPlus cat) -> CatPlus cat infix 9
A bidirectional pattern synonym that matches the (non-terminal) category of an annotated label (of type Maybe cat) and its context (of type cat -> CatPlus cat)). This is can be seen as a wrapper of getCat.

pattern (:#||:) :: (cat, DepMarking) -> (cat -> DepMarking -> CatPlus cat) -> CatPlus cat infix 9
A bidirectional pattern synonym that matches the (non-terminal) category and the grammatical role of an annotated label (of type Maybe (cat, DepMarking)) and its context (of type cat -> DepMarking -> CatPlus cat)). This is can be seen as a wrapper of getCatRule.

Constructors

newNonTerm :: cat -> CatPlus cat

A smart constructor that generates a non-terminal categorial representation without attributes, given a Keyaki category.

Viewers

getCat

Arguments

:: CatPlus cat
An annotated node label

-> (Maybe cat, cat -> CatPlus cat)
Its category and the context.

Extract from an annotated (non-terminal) node label a costate wrt its category, i.e. a pair of its category (of type Maybe cat) and the annotational remainder (aka. the context of the category, an annotated node label CatPlus cat missing a category cat). For annotated terminal node labels, the resulting category is Nothing.

getCatRule

Arguments

:: CatPlus cat
An annotated node label | The pair of its category and its grammatical role, accompanied with their context.

-> (Maybe (cat, DepMarking), cat -> DepMarking -> CatPlus cat)

Extract from an annotated (non-terminal) node label a costate wrt its category and its grammatical rule (i.e. a pair than consists of the pair of its category and its grammatical rule (of type Maybe (cat, DepMarking)) and the annotational remainder (aka. the context of the abovementioned two, of type cat -> DepMarking -> CatPlus cat)). For annotated terminal node labels, the resulting category is Nothing.