KeyakiCategory

Documentation

data KeyakiCategory

Constructors

KeyakiCategory

Fields

catlist :: [String]

iched :: String

sortInfo :: String

createBase :: [String] -> KeyakiCategory

parser :: Parser KeyakiCategory

createFromString :: String -> Either ParseError KeyakiCategory

generate an Keyaki-Category from a string or a stream.

data ParseError

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.