PTDumpable

Description

Provide an type class type of which supports pretty serialization in the munge-trees style, comparable to the Show class in Prelude.

Types

data TreeStyle

TreeStyle specifies the serialization style of trees.

Constructors

OneLine
Each tree is serialized in one line.

Pretty
Each tree is serialized prettily in the same way of munge-trees.

data NodeStyle

NodeStyle speficies the serialized style of node labels.

Constructors

Minimal
Each node is serialized as short as is possible.

Default
Each node is serialized normally.

Full
Each node is serialized with full information.

data Option

Option bundles all the serialization options mentioned above.

Constructors

Option

Fields

treeStyle :: TreeStyle

The serialization style of tree structures.

nodeStyle :: NodeStyle

The serialization style of nodes.

Classes

class Show a => Dumpable a where

Dumpable is a class of types which support the munge-trees style serialization of trees. A tree is serialized with Builder, which needs further conversion to Data.Text or others.

Minimal complete definition

Nothing

Methods

psdDump :: Option -> a -> Builder

Serialize a tree with options.

psdDumpDefault :: a -> Builder

Serialized a tree with the default option.

treeStyle: Pretty

nodeStyle: Default