PTPrintable

Documentation

data TreeStyle

TreeStyle specifies the serialization style of trees.

Constructors

OneLine
Each tree is serialized in one line.

Pretty
Each tree is serialized prettily in the same way of munge-trees.

data NodeStyle

NodeStyle speficies the serialized style of node labels.

Constructors

Minimal
Each node is serialized as short as is possible.

Default
Each node is serialized normally.

Full
Each node is serialized with full information.

data Option

Option bundles all the serialization options mentioned above.

Constructors

Option

Fields

treeStyle :: TreeStyle

The serialization style of tree structures.

nodeStyle :: NodeStyle

The serialization style of nodes.

class Show a => Printable a where

Minimal complete definition

Nothing

Methods

psdPrint :: Option -> a -> String

psdPrintDefault :: a -> String