ABCCategory

Description

Provide an representation of categories of the ABC Grammar. The parser is available at the Parser module.

Types

data ABCCategory

ABCCategory represents ABC categories.

Constructors

Bottom
The Bottom ⊥.

BaseCategory
An atomic catetory.

Fields

name :: Text

The name of the category.

LeftFunctor
A functor category X\C which takes an X to its left as its argument.

Fields

antecedent :: ABCCategory

The argument, corresponding to the X above.

consequence :: ABCCategory

The base, corresponding to the C above.

RightFunctor
A functor category C/X which takes an X to its right as its argument.

Fields

antecedent :: ABCCategory

The argument, corresponding to the X above.

consequence :: ABCCategory

The base, corresponding to the C above.

Subsidiaries

data ABCStatusFC

Constructors

FCLeft Int
Indicating that a left reduction/functional-composition rule is used.

FCRight Int
Indicating that a right reduction/functional-composition rule is used.

Failed
Indicating that the reduction cannot be achieved.

Constants

strBot :: Text

The text representation of the bottom.

Functions

For Composition

(</>)

Arguments

:: ABCCategory
The base category C

-> ABCCategory
The argument category X.

-> ABCCategory
The resulted category C/X.

Compose two categories to form a right functor category.

Example:

>>> (BaseCategory "VP") </> (BaseCategory "NP")
RightFunctor (BaseCategory "NP") (BaseCategory "VP")

makeLeftAdjunct

Arguments

:: ABCCategory
The argument category X.

-> ABCCategory
The resulted category X\X.

Make an endomorphic adjunct X\X from a category X.

Example:

>>> makeLeftAdjunct (BaseCategory "S")
LeftFunctor (BaseCategory "S") (BaseCategory "S")

(<\>)

Arguments

:: ABCCategory
The argument category X.

-> ABCCategory
The base category C

-> ABCCategory
The resulted category X\C.

Compose two categories to form a left functor category.

Example:

>>> (BaseCategory "NP") <\> (BaseCategory "S")
LeftFunctor (BaseCategory "NP") (BaseCategory "S")

makeRightAdjunct

Arguments

:: ABCCategory
The argument category X.

-> ABCCategory
The resulted category X/X.

Make an endomorphic adjunct X/X from a category X.

Example:

>>> makeRightAdjunct (BaseCategory "S")
RightFunctor (BaseCategory "S") (BaseCategory "S")

For Reduction

(<^>)

Arguments

:: ABCCategory
The category to the left.

-> ABCCategory
The category to the right.

-> ABCCategory
The result.

Taking two ABC categories, try a reduction from them.

reduceWithResult

Arguments

:: ABCCategory
The category to the left.

-> ABCCategory
The category to the right.

-> State ABCStatusFC ABCCategory
The result.

Taking two ABC categories, try a reduction from them, leaving the trace of the used rule.

reduceWithLog :: ABCCategory -> ABCCategory -> ABCComment ABCCategory