CatPlus

Description

Provide an representation of categories of the Kainoki Treebank. The parser is available at the Parser module.

Types

data CatPlus cat

Record structure that represents Keyaki categories which augmented with our own additional attributes.

Constructors

Term

Fields

word :: Text

NonTerm

Fields

cat :: cat

index :: Maybe Int

role :: DepMarking

deriv :: Text

scope :: [Int]

covertArgs :: [(Int, cat)]

attrs :: Map Text Text

Bundled Patterns

pattern (:#:) :: cat -> (cat -> CatPlus cat) -> CatPlus cat infix 9

pattern (:#||:) :: (cat, DepMarking) -> (cat -> DepMarking -> CatPlus cat) -> CatPlus cat infix 9

Constructors

newNonTerm :: cat -> CatPlus cat

Smart constructor that generates a non-terminal categorial representation without attributes, given a Keyaki category.

Viewers

getCat :: CatPlus cat -> (Maybe cat, cat -> CatPlus cat)